function that executes a function given as a parameter on each element of an array

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -11,7 +11,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int index;
+	unsigned int index;
 
 	for (index = 0; index < size; index++)
 		(*action)(*(array + index));
